# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lingmo",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/ab-test": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Ab Test Opening",
        "description": "A/B test: generate chapter 1 with multiple writer voices, find optimal.",
        "operationId": "ab_test_opening_api_ab_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ab Test Opening Api Ab Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/final-polish": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Final Polish",
        "description": "Run final full-book polish before publication.",
        "operationId": "final_polish_api_novels__novel_id__final_polish_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Final Polish Api Novels  Novel Id  Final Polish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/polish": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Polish Novel",
        "description": "Polish all generated chapters.",
        "operationId": "polish_novel_api_novels__novel_id__polish_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Polish Novel Api Novels  Novel Id  Polish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/evolve": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Evolve Novel",
        "description": "Run iterative evolution until quality improves or limits are reached.",
        "operationId": "evolve_novel_api_novels__novel_id__evolve_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evolve Novel Api Novels  Novel Id  Evolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/chapters/{chapter_num}/humanize": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Humanize Chapter",
        "description": "Deep-humanize one chapter.",
        "operationId": "humanize_chapter_api_novels__novel_id__chapters__chapter_num__humanize_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          },
          {
            "name": "chapter_num",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chapter Num"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Humanize Chapter Api Novels  Novel Id  Chapters  Chapter Num  Humanize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/chapters/{chapter_num}/revise": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Revise Chapter",
        "description": "Revise a chapter from natural-language critique.",
        "operationId": "revise_chapter_api_novels__novel_id__chapters__chapter_num__revise_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          },
          {
            "name": "chapter_num",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chapter Num"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Revise Chapter Api Novels  Novel Id  Chapters  Chapter Num  Revise Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/revise-opening": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Trigger Revise Opening",
        "description": "Revise the opening chapters using ending knowledge.",
        "operationId": "trigger_revise_opening_api_novels__novel_id__revise_opening_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Revise Opening Api Novels  Novel Id  Revise Opening Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/generate-classic": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Trigger Generate Classic",
        "description": "Generate in classic multi-candidate mode.",
        "operationId": "trigger_generate_classic_api_novels__novel_id__generate_classic_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Generate Classic Api Novels  Novel Id  Generate Classic Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/consistency-score": {
      "get": {
        "tags": [
          "revision"
        ],
        "summary": "Get Consistency Score",
        "description": "Get cross-chapter consistency score.",
        "operationId": "get_consistency_score_api_novels__novel_id__consistency_score_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Consistency Score Api Novels  Novel Id  Consistency Score Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/batch-generate": {
      "post": {
        "tags": [
          "revision"
        ],
        "summary": "Trigger Batch Generate",
        "description": "Generate N chapters sequentially with fixed constraint level for long-run testing.",
        "operationId": "trigger_batch_generate_api_novels__novel_id__batch_generate_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "default": {},
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Batch Generate Api Novels  Novel Id  Batch Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}